Refactor ballet.validation -> ballet.validation.main
