+++ tests/end_to_end/test_validation_end_to_end.py
@@ -18,7 +18,7 @@ from ballet.util import get_enum_values
 from ballet.util.git import make_commit_range, switch_to_new_branch
 from ballet.util.log import logger
 from ballet.util.mod import import_module_at_path, modname_to_relpath
-from ballet.validation import TEST_TYPE_ENV_VAR, BalletTestTypes
+from ballet.validation.main import BalletTestTypes, TEST_TYPE_ENV_VAR
 
 
 def submit_feature(repo, contrib_dir, username, featurename, new_feature_str):
